docs: correct what the README got wrong about the API it describes
Six things, found by checking every documented number, type and signature
against the code:

request_to_get was written as taking nothing, though its own Arguments list
describes the set of EPCs it needs. Its one-line description also said the
properties are sent rather than fetched.

get_historical_cumulative_energy_3 was said to cover six hours. It reads one
minute per point and takes at most ten of them, so it covers ten minutes -
six hours is the figure for _2, which is thirty minutes across twelve points.

get_instantaneous_power was said to return a float. It returns whatever
int.from_bytes gives, which is an int; the annotations said float too and now
say int, which is the narrower and true one.

The async cancellation note still quoted an hour for how long an abandoned
call can hold a worker. Lowering xmit_timeout to 300 made that seven minutes,
and its example still recommended a value below the old default as if it were
short.

lqi and rssi were missing from the list of what AsyncMomonga exposes
read-only.

Checked and correct as they stand: every other default, the 144 s and 15 s
figures, the data-point ranges, the property maps, every signature, and that
no public method or MomongaNeedToReopen subclass is undocumented.

+++ momonga/momonga.py
@@ -636,7 +636,7 @@ class Momonga:
         res = self._request_to_get([req])[0]
         return EchonetDataParser.parse_day_for_historical_data_1(res.edt)
 
-    def get_instantaneous_power(self) -> float:
+    def get_instantaneous_power(self) -> int:
         req = EchonetProperty(EchonetPropertyCode.instantaneous_power)
         res = self._request_to_get([req])[0]
         return EchonetDataParser.parse_instantaneous_power(res.edt)
